+++ packages/components/ui/CHANGELOG.md
@@ -1,5 +1,7 @@
-## `@svizzle/ui` v0.2.1 (next)
+## `@svizzle/ui` v0.3.0 (next)
 
+- add `src/LoadingView.svelte`
+- add `src/MessageView.svelte`
 - updated some dev dependencies
 
 ## `@svizzle/ui` v0.2.0